% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raken.R
\name{rakenfenix}
\alias{rakenfenix}
\title{Create Histogram, boxplot and numeric summary}
\usage{
rakenfenix(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Create Histogram, boxplot and numeric summary
}
